Adds beyondtrust lib, moves functionality into library, shares those functions to two modules
